chore: fix price input resetting when quote received + tidying up
