+++ packages/client/src/client/App/Credit/SellSide/SellSideTradeTicket/SellSideTradeTickerFooter.tsx
@@ -1,5 +1,4 @@
 import { createSignal } from "@react-rxjs/utils"
-import { ThemeName } from "client/theme"
 import {
   customNumberFormatter,
   invertDirection,
@@ -11,7 +10,6 @@ import {
   PASSED_QUOTE_STATE,
   PENDING_WITHOUT_PRICE_QUOTE_STATE,
   QuoteBody,
-  QuoteState,
   RfqState,
 } from "generated/TradingGateway"
 import { FaCheckCircle, FaThumbsDown } from "react-icons/fa"
@@ -21,89 +19,20 @@ import {
   quoteCreditQuote$,
   useCreditRfqDetails,
 } from "services/credit"
-import styled from "styled-components"
 
 import { CreditRfqTimer, isRfqTerminated } from "../../common"
+import { price$, usePrice } from "../sellSideState"
 import { useIsFocused } from "../utils/useIsFocused"
-import { price$, usePrice } from "./SellSideTradeTicketParameters"
-
-const FooterWrapper = styled.div<{ accepted: boolean; missed: boolean }>`
-  display: flex;
-  align-items: center;
-  height: 50px;
-  padding: 8px;
-  border-top: 1px solid ${({ theme }) => theme.primary[3]};
-  background-color: ${({ accepted, missed }) =>
-    accepted
-      ? "rgba(1, 195, 141, 0.1)"
-      : missed
-      ? "rgba(167, 39, 64, 0.15)"
-      : undefined};
-`
-
-export const TimerWrapper = styled.div`
-  display: flex;
-  align-items: center;
-  flex: 1;
-  padding: 0 8px;
-`
-
-const FooterButton = styled.button`
-  padding: 4px 8px;
-  border-radius: 3px;
-  height: 24px;
-  font-size: 11px;
-  &:focus {
-    border-radius: 3px;
-    border: 1px solid #4c76c4 !important;
-  }
-`
-
-const PassButton = styled(FooterButton)<{ disabled: boolean }>`
-  color: ${({ theme }) =>
-    theme.secondary[theme.name === ThemeName.Dark ? "base" : 3]};
-  background-color: ${({ theme }) =>
-    theme.primary[theme.name === ThemeName.Dark ? 2 : 3]};
-  ${({ disabled }) => (disabled ? "opacity: 0.3" : "")}
-`
-
-const SendQuoteButton = styled(FooterButton)<{
-  direction: Direction
-  disabled: boolean
-}>`
-  color: ${({ theme }) => theme.white};
-  background-color: ${({ theme, direction }) =>
-    theme.colors.spectrum.uniqueCollections[direction].darker};
-  ${({ disabled }) => (disabled ? "opacity: 0.3" : "")}
-`
-
-const TradeStatus = styled.div`
-  width: 100%;
-  display: flex;
-  align-items: center;
-  font-size: 11px;
-  font-weight: 700;
-  svg {
-    margin-right: 8px;
-  }
-`
-const Terminated = styled(TradeStatus)`
-  justify-content: center;
-  color: ${({ theme }) => theme.primary[5]};
-`
-const Accepted = styled(TradeStatus)`
-  color: ${({ theme }) => theme.accents.positive.base};
-`
-const Missed = styled(TradeStatus)`
-  color: ${({ theme }) =>
-    theme.accents.negative[
-      theme.name === ThemeName.Light ? "darker" : "medium"
-    ]};
-`
-const TradeDetails = styled.div`
-  font-size: 9px;
-  font-weight: 500px;
-`
+import {
+  Accepted,
+  FooterWrapper,
+  Missed,
+  PassButton,
+  SendQuoteButton,
+  Terminated,
+  TimerWrapper,
+  TradeDetails,
+} from "./SellSideTradeTicketFooter.styles"
 
 const [sendQuoteId$, sendQuote] = createSignal<number>()
