+++ packages/client/src/client/App/Credit/SellSide/SellSideTradeTicket/SellSideTradeTicketCore.tsx
@@ -1,66 +1,18 @@
 import { Loader } from "client/components/Loader"
 import { invertDirection } from "client/utils"
-import {
-  PENDING_WITHOUT_PRICE_QUOTE_STATE,
-  QuoteBody,
-} from "generated/TradingGateway"
+import { PENDING_WITHOUT_PRICE_QUOTE_STATE } from "generated/TradingGateway"
 import { useCreditRfqDetails } from "services/credit"
-import styled from "styled-components"
 
 import { getSellSideQuoteState, SellSideQuoteState } from "../sellSideState"
-import { getSellSideStatusColor } from "../utils"
 import { SellSideTradeTicketFooter } from "./SellSideTradeTickerFooter"
+import {
+  Banner,
+  Diamond,
+  SellSideTradeTicketInnerWrapper,
+  SellSideTradeTicketWrapper,
+} from "./SellSideTradeTicketCore.styles"
 import { SellSideTradeTicketHeader } from "./SellSideTradeTicketHeader"
 import { SellSideTradeTicketParameters } from "./SellSideTradeTicketParameters"
-
-const SellSideTradeTicketWrapper = styled.div`
-  display: flex;
-  flex-direction: column;
-  padding: 8px;
-  gap: 8px;
-  height: 100%;
-  background-color: ${({ theme }) => theme.core.lightBackground};
-`
-
-const SellSideTradeTicketInnerWrapper = styled.div`
-  display: flex;
-  flex-direction: column;
-  border: 1px solid #686d74;
-  border-radius: 2px;
-  height: 100%;
-  background-color: ${({ theme }) => theme.core.lightBackground};
-`
-const Banner = styled.div`
-  display: flex;
-  align-items: center;
-  padding: 8px;
-  gap: 8px;
-  color: ${({ theme }) => theme.textColor};
-  font-size: 11px;
-  line-height: 16px;
-  font-weight: 500;
-`
-
-const Diamond = styled.div<{ state: SellSideQuoteState }>`
-  width: 0;
-  height: 0;
-  border: 4px solid transparent;
-  border-bottom-color: ${({ theme, state }) =>
-    getSellSideStatusColor(state, theme)};
-  position: relative;
-  top: -4px;
-  &:after {
-    content: "";
-    position: absolute;
-    left: -4px;
-    top: 4px;
-    width: 0;
-    height: 0;
-    border: 4px solid transparent;
-    border-top-color: ${({ theme, state }) =>
-      getSellSideStatusColor(state, theme)};
-  }
-`
 
 const ADAPTIVE_BUYSIDE_NAME = "Adaptive Asset Management"
