+++ packages/client/src/client/App/Credit/SellSide/SellSideTradeTicket/SellSideTradeTicketParameters.tsx
@@ -1,118 +1,24 @@
-import { bind } from "@react-rxjs/core"
-import { createSignal } from "@react-rxjs/utils"
-import { ThemeName } from "client/theme"
-import {
-  DECIMAL_SEPARATOR,
-  DECIMAL_SEPARATOR_REGEXP,
-  THOUSANDS_SEPARATOR_REGEXP,
-  truncatedDecimalNumberFormatter,
-} from "client/utils"
+import { truncatedDecimalNumberFormatter } from "client/utils"
 import { QuoteBody, RfqState } from "generated/TradingGateway"
-import {
-  AcceptedQuoteState,
-  PendingWithPriceQuoteState,
-} from "generated/TradingGateway"
+import { PendingWithPriceQuoteState } from "generated/TradingGateway"
 import { useEffect, useRef } from "react"
-import { merge } from "rxjs"
-import { filter, map } from "rxjs/operators"
-import styled from "styled-components"
 
 import {
   getSellSideQuoteState,
-  selectedRfqId$,
   SellSideQuoteState,
+  setPrice,
+  usePrice,
 } from "../sellSideState"
 import { setFocused } from "../utils/useIsFocused"
-
-const ParametersWrapper = styled.div`
-  padding: 8px;
-  overflow-y: auto;
-  font-size: 12px;
-  font-weight: 500;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-`
-const ParameterLabel = styled.div`
-  color: ${({ theme }) => theme.secondary[5]};
-  margin-bottom: 4px;
-`
-
-const ParameterValue = styled.div`
-  display: flex;
-  align-items: center;
-  height: 24px;
-  color: ${({ theme }) =>
-    theme.secondary[theme.name === ThemeName.Dark ? "base" : 1]};
-`
-
-const PendingPrice = styled.span`
-  margin-right: 6px;
-  color: ${({ theme }) => theme.accents.primary.base};
-`
-
-export const ParameterInput = styled.input`
-  outline: none;
-  height: 24px;
-  width: 100%;
-  padding: 4px;
-  color: ${({ theme }) =>
-    theme.secondary[theme.name === ThemeName.Dark ? "base" : 1]};
-  background-color: ${({ theme }) => theme.core.darkBackground};
-  border: 1px solid ${({ theme }) => theme.colors.light.primary[3]};
-  border-radius: 3px;
-  &:focus {
-    outline: none !important;
-    border-color: ${({ theme }) => theme.accents.primary.darker};
-  }
-  ¯ &:disabled {
-    opacity: 0.3;
-  }
-  &.is-invalid {
-    border-color: ${({ theme }) => theme.accents.negative.base};
-  }
-`
+import {
+  ParameterInput,
+  ParameterLabel,
+  ParametersWrapper,
+  ParameterValue,
+  PendingPrice,
+} from "./SellSideTradeTicketParameters.styles"
 
 const formatter = truncatedDecimalNumberFormatter(4)
-
-const filterRegExp = new RegExp(THOUSANDS_SEPARATOR_REGEXP, "g")
-const decimalRegExp = new RegExp(DECIMAL_SEPARATOR_REGEXP, "g")
-
-const [rawPrice$, setPrice] = createSignal<string>()
-export const [usePrice, price$] = bind(
-  merge(
-    selectedRfqId$.pipe(map(() => ({ value: 0, inputValue: "" }))),
-    rawPrice$.pipe(
-      map((rawVal) => {
-        const lastChar = rawVal.slice(-1).toLowerCase()
-        const cleanedInput = rawVal
-          .replace(filterRegExp, "")
-          .replace(decimalRegExp, ".")
-
-        const inputQuantityAsNumber = Math.abs(Number(cleanedInput))
-
-        // numeric value could be NaN at this stage
-
-        const truncated = formatter(inputQuantityAsNumber)
-
-        const value = Number(
-          truncated.replace(filterRegExp, "").replace(decimalRegExp, "."),
-        )
-
-        return {
-          value,
-          inputValue:
-            value === 0
-              ? ""
-              : formatter(value) +
-                (lastChar === DECIMAL_SEPARATOR ? DECIMAL_SEPARATOR : ""),
-        }
-      }),
-      filter(({ value }) => !Number.isNaN(value)),
-    ),
-  ),
-  { value: 0, inputValue: "" },
-)
 
 interface SellSideTradeTicketParametersProps {
   selectedRfqId: number
@@ -172,9 +78,7 @@ export const SellSideTradeTicketParameters = ({
           </ParameterValue>
         ) : (
           <ParameterValue>
-            {quote?.state && "payload" in quote.state
-              ? quote.state.payload
-              : ""}
+            {quote?.state && "payload" in quote.state && quote.state.payload}
           </ParameterValue>
         )}
       </div>
